refactor(diffusion): move train-pair expansion into algorithm hooks
Keep rollout orchestration generic while allowing NFT to provide its own sample expansion without a converter class.
