+++ miles/ray/data_conversion_hub/__init__.py
@@ -1,0 +1,1 @@
+"""Rollout-to-train-pair conversion hooks for diffusion algorithms."""
