+++ miles/ray/rollout.py
@@ -13,6 +13,9 @@ from ray.util.scheduling_strategies import PlacementGroupSchedulingStrategy
 from sglang.srt.constants import GPU_MEMORY_TYPE_WEIGHTS
 
 from miles.backends.sglang_diffusion_utils.sglang_diffusion_engine import SGLangDiffusionEngine
+from miles.ray.data_conversion_hub.flow_grpo import (
+    expand_samples_to_train_pairs as flow_grpo_expand_samples_to_train_pairs,
+)
 from miles.rollout.base_types import call_rollout_fn
 from miles.rollout.rm_hub.core import set_reward_placement_group
 from miles.utils import tracking_utils
@@ -26,11 +29,7 @@ from miles.utils.misc import load_function
 from miles.utils.ray_utils import Box
 from miles.utils.timer import timer
 from miles.utils.tracking_utils import init_tracking
-from miles.utils.train_data_utils import (
-    TrainDataDPSplitter,
-    reorder_train_pairs_for_tiling,
-    resolve_train_data_converter,
-)
+from miles.utils.train_data_utils import TrainDataDPSplitter, reorder_train_pairs_for_tiling
 from miles.utils.train_metric_utils import log_perf_data_raw
 from miles.utils.types import Sample
 
@@ -81,9 +80,11 @@ class RolloutManager:
             if self.args.custom_convert_samples_to_train_data_path is not None
             else None
         )
-        # Algorithm-specific converters share the default reward post-process path.
-        # Full convert overrides use --custom-convert-samples-to-train-data-path instead.
-        self.train_data_converter = resolve_train_data_converter(self.args)
+        self.custom_expand_samples_to_train_pairs_func = (
+            load_function(self.args.custom_expand_samples_to_train_pairs_path)
+            if self.args.custom_expand_samples_to_train_pairs_path is not None
+            else None
+        )
         self.train_data_dp_splitter = TrainDataDPSplitter()
         logger.info(f"import {self.args.rollout_function_path} as generate_rollout function.")
         logger.info(f"import {self.args.eval_function_path} as eval_generate_rollout function.")
@@ -341,9 +342,27 @@ class RolloutManager:
         if self.custom_reward_post_process_func is not None:
             return self.custom_reward_post_process_func(self.args, samples)
 
-        from miles.backends.fsdp_utils.loss_hub.advantages import grpo_normalize_rewards
+        raw_rewards = [sample.get_reward_value(self.args) for sample in samples]
 
-        return grpo_normalize_rewards(self.args, samples)
+        # --globalize-reward-mean / --globalize-reward-std are orthogonal.
+        rewards_flat = torch.tensor(raw_rewards, dtype=torch.float)
+        rewards = rewards_flat.view(-1, self.args.n_samples_per_prompt)
+
+        if self.args.globalize_reward_mean:
+            mean = rewards_flat.mean()
+        else:
+            mean = rewards.mean(dim=-1, keepdim=True)
+        rewards = rewards - mean
+
+        if self.args.grpo_std_normalization:
+            if self.args.globalize_reward_std:
+                std = rewards_flat.std()
+            else:
+                std = rewards.std(dim=-1, keepdim=True)
+            # matches flow_grpo's `+ 1e-4` in both stat_tracking branches
+            rewards = rewards / (std + 1e-4)
+
+        return raw_rewards, rewards.flatten().tolist()
 
     def _convert_samples_to_train_data(self, samples: list[Sample] | list[list[Sample]]):
         """
@@ -395,7 +414,9 @@ class RolloutManager:
                 reward_key=self.args.reward_key,
             )
 
-        return self.train_data_converter.convert_samples(samples, rewards, raw_rewards)
+        if self.custom_expand_samples_to_train_pairs_func is not None:
+            return self.custom_expand_samples_to_train_pairs_func(self.args, samples, rewards, raw_rewards)
+        return flow_grpo_expand_samples_to_train_pairs(self.args, samples, rewards, raw_rewards)
 
     def _log_images(
         self,
